WIP: Initial Count Down View Work
Implemented a custom count down view which accepts start and end time values and an optional boolean for either enabled / disabling pulsation. Implemented some simple animations which scales the text and fades it in and out at specified time intervals. Implemented a bunch of callbacks in an interface for common callback events such as start, stop, pause, tick, resume and finished.

Todo:

1) Add the ability to changed the time remaining
2) Resolve an issue where XML start and end times are not being adhered to